# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TournaX Gaming Tournament API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface RegisterRequest {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  player: "player",
  host: "host",
  admin: "admin",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  pending: "pending",
  active: "active",
  banned: "banned",
} as const;

export interface User {
  id: number;
  email: string;
  name?: string;
  handle?: string;
  avatar?: string;
  game?: string;
  gameUid?: string;
  role: UserRole;
  balance: number;
  status: UserStatus;
  followersCount?: number;
  followingCount?: number;
  profileSetup: boolean;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface SetupProfileRequest {
  avatar: string;
  game: string;
  ign: string;
  handle: string;
  gameUid: string;
}

export interface UpdateProfileRequest {
  name?: string;
  handle?: string;
  avatar?: string;
}

export type MatchStatus = (typeof MatchStatus)[keyof typeof MatchStatus];

export const MatchStatus = {
  upcoming: "upcoming",
  live: "live",
  completed: "completed",
} as const;

export interface Match {
  id: number;
  code: string;
  game: string;
  mode: string;
  teamSize: number;
  entryFee: number;
  prizePool: number;
  startTime: string;
  status: MatchStatus;
  slots: number;
  filledSlots: number;
  hostHandle?: string;
  hostName?: string;
  hostAvatar?: string;
  hostFollowers?: number;
  isFollowingHost?: boolean;
  isJoined?: boolean;
  roomId?: string;
  roomPassword?: string;
  roomReleased?: boolean;
  hostId: number;
}

export interface CreateMatchRequest {
  game: string;
  mode: string;
  teamSize: number;
  entryFee: number;
  slots: number;
  startTime: string;
}

export interface JoinPlayer {
  ign: string;
  uid: string;
}

export interface JoinMatchRequest {
  teamName?: string;
  players: JoinPlayer[];
}

export interface ParticipantPlayer {
  ign: string;
  uid: string;
  position: number;
}

export interface MatchParticipant {
  id: number;
  teamName?: string;
  teamNumber: number;
  players: ParticipantPlayer[];
}

export interface RoomCredentialsRequest {
  roomId: string;
  roomPassword: string;
}

export interface MyMatchesResponse {
  participated: Match[];
  history: Match[];
}

export interface UserProfile {
  id: number;
  name?: string;
  handle: string;
  avatar?: string;
  role?: string;
  followersCount: number;
  followingCount: number;
  rating?: number;
  matchesCount: number;
  isFollowing: boolean;
  upcomingMatches: Match[];
  activeMatches: Match[];
}

export interface ExploreResponse {
  recommendedHosts: UserProfile[];
  mostActivePlayers: UserProfile[];
}

export interface SquadMember {
  id: number;
  name: string;
  uid: string;
}

export interface AddSquadMemberRequest {
  name: string;
  uid: string;
}

export type WalletTransactionStatus =
  (typeof WalletTransactionStatus)[keyof typeof WalletTransactionStatus];

export const WalletTransactionStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface WalletTransaction {
  id: number;
  amount: number;
  status: WalletTransactionStatus;
  createdAt: string;
  note?: string;
}

export interface WalletInfo {
  balance: number;
  upiId?: string;
  addBalanceHistory: WalletTransaction[];
  withdrawalHistory: WalletTransaction[];
}

export interface AddBalanceRequest {
  utrNumber: string;
  amount: number;
  receiptUrl?: string;
}

export interface WithdrawRequest {
  amount: number;
  upiId: string;
}

export interface Notification {
  id: number;
  type: string;
  message: string;
  read: boolean;
  createdAt: string;
}

export interface Message {
  id: number;
  fromHandle: string;
  fromName?: string;
  content: string;
  createdAt: string;
  read: boolean;
}

export interface SendMessageRequest {
  toHandle: string;
  content: string;
}

export interface ComplaintRequest {
  subject: string;
  description: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name?: string;
  role: string;
}

export interface AdminDashboard {
  totalPlayers: number;
  activePlayers: number;
  pendingKyc: number;
  hosts: number;
  totalMatches: number;
  liveNow: number;
  pendingWithdrawals: number;
  totalRevenue: number;
  platformFees: number;
  complaintsCount: number;
  adminList: AdminUser[];
  hostList: AdminUser[];
}

export type AdminPlayerStatus =
  (typeof AdminPlayerStatus)[keyof typeof AdminPlayerStatus];

export const AdminPlayerStatus = {
  pending: "pending",
  active: "active",
  banned: "banned",
} as const;

export interface AdminPlayer {
  id: number;
  name?: string;
  email: string;
  uid?: string;
  handle?: string;
  balance: number;
  status: AdminPlayerStatus;
  matchesPlayed: number;
}

export type FinanceRequestStatus =
  (typeof FinanceRequestStatus)[keyof typeof FinanceRequestStatus];

export const FinanceRequestStatus = {
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface FinanceRequest {
  id: number;
  userId: number;
  userName?: string;
  userEmail?: string;
  amount: number;
  status: FinanceRequestStatus;
  utrNumber?: string;
  upiId?: string;
  createdAt: string;
}

export interface Complaint {
  id: number;
  userId: number;
  userName?: string;
  subject: string;
  description: string;
  createdAt: string;
}

export interface CreateHostRequest {
  email: string;
  password: string;
  name: string;
}

export interface AdminAddBalanceRequest {
  amount: number;
}

export type ListMatchesParams = {
  search?: string;
  status?: ListMatchesStatus;
};

export type ListMatchesStatus =
  (typeof ListMatchesStatus)[keyof typeof ListMatchesStatus];

export const ListMatchesStatus = {
  upcoming: "upcoming",
  live: "live",
  completed: "completed",
} as const;

export type ExploreUsersParams = {
  search?: string;
};

export type AdminListPlayersParams = {
  search?: string;
  status?: AdminListPlayersStatus;
};

export type AdminListPlayersStatus =
  (typeof AdminListPlayersStatus)[keyof typeof AdminListPlayersStatus];

export const AdminListPlayersStatus = {
  all: "all",
  verified: "verified",
  pending: "pending",
  banned: "banned",
} as const;

export type AdminListAddRequestsParams = {
  status?: AdminListAddRequestsStatus;
};

export type AdminListAddRequestsStatus =
  (typeof AdminListAddRequestsStatus)[keyof typeof AdminListAddRequestsStatus];

export const AdminListAddRequestsStatus = {
  all: "all",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type AdminListWithdrawalsParams = {
  status?: AdminListWithdrawalsStatus;
};

export type AdminListWithdrawalsStatus =
  (typeof AdminListWithdrawalsStatus)[keyof typeof AdminListWithdrawalsStatus];

export const AdminListWithdrawalsStatus = {
  all: "all",
  pending: "pending",
  approved: "approved",
  rejected: "rejected",
} as const;

export type AdminListMatchesParams = {
  status?: AdminListMatchesStatus;
};

export type AdminListMatchesStatus =
  (typeof AdminListMatchesStatus)[keyof typeof AdminListMatchesStatus];

export const AdminListMatchesStatus = {
  all: "all",
  upcoming: "upcoming",
  live: "live",
  completed: "completed",
} as const;