# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TournaX Gaming Tournament API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: Register success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/setup-profile:
    post:
      operationId: setupProfile
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetupProfileRequest"
      responses:
        "200":
          description: Profile setup success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /admin/players:
    get:
      operationId: adminListPlayers
      tags: [admin]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [all, verified, pending, banned]
      responses:
        "200":
          description: Players list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminPlayer"
  /admin/players/{id}/verify:
    post:
      operationId: verifyPlayer
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/players/{id}/ban:
    post:
      operationId: banPlayer
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Banned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/players/{id}/add-balance:
    post:
      operationId: adminAddBalance
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminAddBalanceRequest"
      responses:
        "200":
          description: Balance added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/finance/add-requests:
    get:
      operationId: adminListAddRequests
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, approved, rejected]
      responses:
        "200":
          description: Add balance requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceRequest"
  /admin/finance/add-requests/{id}/approve:
    post:
      operationId: approveAddRequest
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/finance/add-requests/{id}/reject:
    post:
      operationId: rejectAddRequest
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/finance/withdrawals:
    get:
      operationId: adminListWithdrawals
      tags: [admin]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [all, pending, approved, rejected]
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FinanceRequest"
  /admin/finance/withdrawals/{id}/approve:
    post:
      operationId: approveWithdrawal
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/finance/withdrawals/{id}/reject:
    post:
      operationId: rejectWithdrawal
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/create-host:
    post:
      operationId: adminCreateHost
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostRequest"
      responses:
        "200":
          description: Host created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/create-admin:
    post:
      operationId: adminCreateAdmin
      tags: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHostRequest"
      responses:
        "200":
          description: Admin created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required: [user, token]
    SetupProfileRequest:
      type: object
      properties:
        avatar:
          type: string
        game:
          type: string
        ign:
          type: string
        handle:
          type: string
        gameUid:
          type: string
      required: [avatar, game, ign, handle, gameUid]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        handle:
          type: string
        avatar:
          type: string
        game:
          type: string
        gameUid:
          type: string
        role:
          type: string
          enum: [player, host, admin]
        balance:
          type: number
        status:
          type: string
          enum: [pending, active, banned]
        followersCount:
          type: integer
        followingCount:
          type: integer
        profileSetup:
          type: boolean
      required: [id, email, role, balance, status, profileSetup]
...
    AdminPlayer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        uid:
          type: string
        handle:
          type: string
        balance:
          type: number
        status:
          type: string
          enum: [pending, active, banned]
        matchesPlayed:
          type: integer
      required: [id, email, balance, status, matchesPlayed]
    FinanceRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        userEmail:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [pending, approved, rejected]
        utrNumber:
          type: string
        upiId:
          type: string
        createdAt:
          type: string
      required: [id, userId, amount, status, createdAt]
...
    CreateHostRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
      required: [email, password, name]
    AdminAddBalanceRequest:
      type: object
      properties:
        amount:
          type: number
      required: [amount]