# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TournaX Gaming Tournament API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";
import type { UserStatus } from "./userStatus";

export interface User {
  id: number;
  email: string;
  name?: string;
  handle?: string;
  avatar?: string;
  game?: string;
  gameUid?: string;
  role: UserRole;
  balance: number;
  status: UserStatus;
  followersCount?: number;
  followingCount?: number;
  profileSetup: boolean;
}